+++ src/controller/compat.rs
@@ -5,11 +5,12 @@ use std::io::Read;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use gbx::{SCRIPT_API_VERSION, SERVER_API_VERSION};
-
 use crate::config::{Config, BLACKLIST_FILE, MAX_GHOST_REPLAY_RANK, VERSION};
 use crate::database::{Database, Map, MapEvidence};
-use crate::ingame::{MapInfo, ModeInfo, ModeOptions, Server, ServerInfo, ServerOptions};
+use crate::ingame::{
+    MapInfo, ModeInfo, ModeOptions, Server, ServerInfo, ServerOptions, SCRIPT_API_VERSION,
+    SERVER_API_VERSION,
+};
 use crate::network::exchange_id;
 
 /// Runs everything that needs to run at startup.