+++ src/message.rs
@@ -99,20 +99,20 @@ impl ServerMessage<'_> {
             NewPlayerList(PlayerDiff::AddPlayer(info))
             | NewPlayerList(PlayerDiff::AddSpectator(info))
             | NewPlayerList(PlayerDiff::AddPureSpectator(info)) => Some(Joining {
-                nick_name: &info.nick_name,
+                nick_name: &info.nick_name.formatted,
             }),
 
             NewPlayerList(PlayerDiff::RemovePlayer(info))
             | NewPlayerList(PlayerDiff::RemoveSpectator(info))
             | NewPlayerList(PlayerDiff::RemovePureSpectator(info)) => Some(Leaving {
-                nick_name: &info.nick_name,
+                nick_name: &info.nick_name.formatted,
             }),
 
             BeginIntro {
                 loaded_map,
                 is_restart: false,
             } => Some(CurrentMap {
-                name: &loaded_map.name,
+                name: &loaded_map.name.formatted,
                 author: &loaded_map.author_login,
             }),
 
@@ -130,7 +130,7 @@ impl ServerMessage<'_> {
                         ..
                     },
             } if *pos_gained > 0 && *new_pos <= MAX_ANNOUNCED_RECORD => Some(TopRecord {
-                player_nick_name: &new_record.player_nick_name,
+                player_nick_name: &new_record.player_nick_name.formatted,
                 new_map_rank: *new_pos,
                 millis: new_record.millis as usize,
             }),
@@ -146,20 +146,24 @@ impl ServerMessage<'_> {
                     },
             } if *pos_gained == 0 && *diff < 0 && *new_pos <= MAX_ANNOUNCED_RECORD_IMPROVEMENT => {
                 Some(TopRecordImproved {
-                    player_nick_name: &new_record.player_nick_name,
+                    player_nick_name: &new_record.player_nick_name.formatted,
                     map_rank: *new_pos,
                     millis: new_record.millis as usize,
                 })
             }
 
             NewPlaylist(PlaylistDiff::AppendNew(map)) => Some(NewMap {
-                name: &map.name,
+                name: &map.name.formatted,
                 author: &map.author_login,
             }),
 
-            NewPlaylist(PlaylistDiff::Append(map)) => Some(AddedMap { name: &map.name }),
+            NewPlaylist(PlaylistDiff::Append(map)) => Some(AddedMap {
+                name: &map.name.formatted,
+            }),
 
-            NewPlaylist(PlaylistDiff::Remove { map, .. }) => Some(RemovedMap { name: &map.name }),
+            NewPlaylist(PlaylistDiff::Remove { map, .. }) => Some(RemovedMap {
+                name: &map.name.formatted,
+            }),
 
             NewServerRanking(ServerRankingDiff { diffs, .. }) => {
                 let mut top_ranks: Vec<TopRankMessage> = diffs
@@ -167,7 +171,7 @@ impl ServerMessage<'_> {
                     .filter_map(|diff| {
                         if diff.gained_pos > 0 && diff.new_pos <= MAX_ANNOUNCED_RANK {
                             Some(TopRankMessage {
-                                nick_name: &diff.player_nick_name,
+                                nick_name: &diff.player_nick_name.formatted,
                                 rank: diff.new_pos,
                             })
                         } else {